fix(journey): address review — shared CP mutex + rate limit on both paths
Review fixes for PR #946:
- Extract serializedCPAward to shared singleton (cpAwardLock.ts)
- Fix queue processing: cpAwardInProgress stays true while draining queue
- Add client-side rate limit guard to momentService.createMoment()
- Both moment creation paths now share the same mutex + rate limit

Note: client-side guards are a stopgap. DB-level FOR UPDATE lock in
award_consciousness_points RPC is the definitive fix (tracked in #936).
